feat(app.js): enable parsing of form data in request body
uses middleware express.urlencoded(). this is required for POST and PUT requests for when the data comes _back_ to recognise incoming Request Object as strings or arrays (by default) but the extended: true within the method call helps also extend this to any type (incl objects). even though I might not use _forms_, I'll probably need this for PUT and POST requests