feat: enable css to be used to style views
not able to have modular css stylesheets for each ejs. instead, AFAIK, i need to set up the 'public' directory in app.js using app.use (for serving static files, like css stylesheets and images), and then add it to each of the ejs files. this was based on stackoverflow answer here: https://stackoverflow.com/a/18633827